Make wco into multi-src
