+++ lib-multisrc/wco/build.gradle.kts
@@ -1,0 +1,6 @@
+baseKmkVersionCode = 0
+plugins {
+    id("lib-multisrc")
+}
+
+baseVersionCode = 1
